feat: Add module and main Go file for variables
This commit adds a new module and a main Go file to implement a program that demonstrates the usage of variables in Go. It defines a constant variable and prints its value along with the value of a string variable. The commit also includes some explanatory comments for better understanding.